requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
fhir.resources>=7.1.0
pytest==7.4.3
requests==2.31.0
pytest-cov==4.1.0
